+++ @kiva/kv-components/tests/unit/specs/components/KvTextInput.spec.js
@@ -28,7 +28,7 @@ describe('KvTextInput', () => {
 		const textInputEl = getByRole('textbox');
 
 		expect(textInputEl.value).toEqual('');
-		await userEvent.type(textInputEl, 'abc 123');
+		userEvent.type(textInputEl, 'abc 123');
 		expect(textInputEl.value).toEqual('abc 123');
 	});
 
@@ -47,6 +47,81 @@ describe('KvTextInput', () => {
 		expect(textInputEl.value).toEqual('');
 	});
 
+	it('works with v-model', async () => {
+		const TestComponent = {
+			template:
+				`<div>
+					<label for="text-input">Text input</label>
+					<KvTextInput v-model="textValue" id="text-input" />
+					<button @click="textValue = 'abc'">reset</button>
+					<span>The text value is {{ textValue }}</span>
+				</div>`,
+			components: { KvTextInput },
+			data: () => ({ textValue: 'abc' }),
+		};
+		const { getByRole, getByText } = render(TestComponent);
+		const textInputEl = getByRole('textbox');
+
+		// Check that the value is 'abc' initially
+		expect(getByText('The text value is abc')).toBeDefined();
+		expect(textInputEl.value).toEqual('abc');
+
+		// Type 'def' in the text input and expect the value to be 'abcdef' now
+		await userEvent.type(textInputEl, 'def');
+		expect(getByText('The text value is abcdef')).toBeDefined();
+		expect(textInputEl.value).toEqual('abcdef');
+
+		// Click the reset button and expect the value to be 'abc' again
+		await fireEvent.click(getByText('reset'));
+		expect(getByText('The text value is abc')).toBeDefined();
+		expect(textInputEl.value).toEqual('abc');
+	});
+
+	it('applies parent event listeners to the input element', async () => {
+		const onInput = jest.fn();
+		const TestComponent = {
+			template: '<KvTextInput id="test" @input="onInput" />',
+			components: { KvTextInput },
+			methods: { onInput },
+		};
+		const { getByRole } = render(TestComponent);
+
+		const textInputEl = getByRole('textbox');
+		await userEvent.type(textInputEl, 'a');
+		expect(onInput.mock.calls.length).toBe(1);
+	});
+
+	it('applies parent attributes to the input element', async () => {
+		const TestComponent = {
+			template: '<KvTextInput id="test" name="test-text-input" />',
+			components: { KvTextInput },
+		};
+		const { getByRole } = render(TestComponent);
+
+		const textInputEl = getByRole('textbox');
+		expect(textInputEl.name).toBe('test-text-input');
+	});
+
+	it('applies parent styles to the root element', async () => {
+		const TestComponent = {
+			template: '<KvTextInput id="test" style="padding-top:1234px" />',
+			components: { KvTextInput },
+		};
+		const { container } = render(TestComponent);
+
+		expect(container.firstChild.style.paddingTop).toEqual('1234px');
+	});
+
+	it('applies parent classes to the root element', async () => {
+		const TestComponent = {
+			template: '<KvTextInput id="test" class="test-class" />',
+			components: { KvTextInput },
+		};
+		const { container } = render(TestComponent);
+
+		expect(container.firstChild.classList).toContain('test-class');
+	});
+
 	it('has no automated accessibility violations', async () => {
 		const { container } = render(KvTextInputTemplate);
 		const results = await axe(container);
@@ -55,7 +130,7 @@ describe('KvTextInput', () => {
 	});
 
 	it('clear button cleans the input value', async () => {
-		const { getByRole } = render(KvTextInput, {
+		const { getByRole, findByRole, queryByRole } = render(KvTextInput, {
 			props: {
 				canClear: true,
 				valid: true,
@@ -63,12 +138,16 @@ describe('KvTextInput', () => {
 			},
 		});
 		const textInputEl = getByRole('textbox');
+
 		expect(textInputEl.value).toEqual('');
-		await userEvent.type(textInputEl, 'abc 123');
+		expect(queryByRole('button')).toBeNull();
+
+		userEvent.type(textInputEl, 'abc 123');
 		expect(textInputEl.value).toEqual('abc 123');
-		const buttonInputEl = getByRole('button');
-		expect(buttonInputEl).toBeDefined();
-		await fireEvent.click(buttonInputEl);
+		const buttonEl = await findByRole('button');
+		expect(buttonEl).toBeDefined();
+
+		await fireEvent.click(buttonEl);
 		expect(textInputEl.value).toEqual('');
 	});
 });