+++ @kiva/kv-components/vue/.storybook/main.cjs
@@ -1,0 +1,25 @@
+const path = require('path');
+
+module.exports = {
+  stories: [
+    '../stories/Styleguide.stories.js', // show the base styleguide first
+    '../stories/**/*.stories.mdx',
+    '../stories/**/*.stories.@(js|jsx|ts|tsx)'
+  ],
+  addons: [
+    '@storybook/addon-links',
+    '@storybook/addon-essentials',
+    '@storybook/addon-a11y',
+    '@storybook/addon-storysource',
+    'storybook-dark-mode',
+    {
+    name: '@storybook/addon-postcss',
+    options: {
+      postcssLoaderOptions: {
+        implementation: require('postcss'),
+      },
+    },
+  },
+],
+  "framework": "@storybook/vue"
+}