+++ @kiva/kv-tokens/configs/tailwind.config.cjs
@@ -1,9 +1,9 @@
 const plugin = require('tailwindcss/plugin');
 const typographyPlugin = require('@tailwindcss/typography');
-const kivaTypography = require('./kivaTypography');
-const { defaultTheme, buildColorChoices } = require('./kivaColors');
+const kivaTypography = require('./kivaTypography.cjs');
+const { defaultTheme, buildColorChoices } = require('./kivaColors.cjs');
 const designtokens = require('../primitives.json');
-const { rem } = require('./util');
+const { rem } = require('./util.cjs');
 
 const {
 	fonts,
@@ -18,8 +18,7 @@ const {
 } = designtokens;
 
 module.exports = {
-	purge: [],
-	darkMode: false, // or 'media' or 'class'
+	content: ['./**.*.js'],
 	prefix: 'tw-', // prefixes all tailwinds classes with tw. e.g., 'tw-flex tw-mb-2'
 	corePlugins: {
 		boxShadow: false,
@@ -170,8 +169,11 @@ module.exports = {
 				body: {
 					...textStyles.textBase,
 					color: textBaseColor,
+					'-webkit-font-smoothing': 'antialiased',
+					'-moz-osx-font-smoothing': 'grayscale',
+					'text-rendering': 'optimizeLegibility',
 				},
-				button: { fontWeight: fontWeights.book },
+				button: { fontWeight: 'inherit' },
 				h1: textStyles.textH1,
 				h2: textStyles.textH2,
 				h3: textStyles.textH3,