+++ @kiva/kv-tokens/CHANGELOG.md
@@ -3,6 +3,61 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0](https://github.com/kiva/kv-ui-elements/compare/@kiva/kv-tokens@1.3.3...@kiva/kv-tokens@2.0.0) (2022-02-03)
+
+
+* feat!: upgrade Tailwind to v3 ([18c1653](https://github.com/kiva/kv-ui-elements/commit/18c1653565eb42b8abfaaba6b3712173ff341850))
+
+
+### BREAKING CHANGES
+
+* Tailwind dependency is now v3
+
+
+
+
+
+## [1.3.3](https://github.com/kiva/kv-ui-elements/compare/@kiva/kv-tokens@1.3.2...@kiva/kv-tokens@1.3.3) (2022-02-03)
+
+**Note:** Version bump only for package @kiva/kv-tokens
+
+
+
+
+
+## [1.3.2](https://github.com/kiva/kv-ui-elements/compare/@kiva/kv-tokens@1.3.1...@kiva/kv-tokens@1.3.2) (2022-02-03)
+
+
+### Bug Fixes
+
+* fix empty content error ([ae08e8a](https://github.com/kiva/kv-ui-elements/commit/ae08e8a8edf27917a0a0d398b27b303f027cb887))
+
+
+
+
+
+## [1.3.1](https://github.com/kiva/kv-ui-elements/compare/@kiva/kv-tokens@1.3.0...@kiva/kv-tokens@1.3.1) (2022-01-11)
+
+
+### Bug Fixes
+
+* allow buttons to inherit font weights ([593b28f](https://github.com/kiva/kv-ui-elements/commit/593b28f5f2b23ad38b893b64ac866e8287f12479))
+
+
+
+
+
+# [1.3.0](https://github.com/kiva/kv-ui-elements/compare/@kiva/kv-tokens@1.2.1...@kiva/kv-tokens@1.3.0) (2021-12-23)
+
+
+### Features
+
+* match previous font smoothing settings ([e6c3edf](https://github.com/kiva/kv-ui-elements/commit/e6c3edf8f696ccd496607ab40b6874a3e63429ea))
+
+
+
+
+
 ## [1.2.1](https://github.com/kiva/kv-ui-elements/compare/@kiva/kv-tokens@1.2.0...@kiva/kv-tokens@1.2.1) (2021-11-24)
 
 **Note:** Version bump only for package @kiva/kv-tokens